some updates

+++ library/Nano/Db/mysql.php
@@ -9,6 +9,9 @@ class Nano_Db_mysql {
 	public static function clean(Nano_Db $db) {
 		$tables = $db->query('show tables', PDO::FETCH_NUM);
 		foreach ($tables as $row) {
+			if (Nano_Migrate::VERSION_TABLE === $row[0]) {
+				continue;
+			}
 			$db->query('truncate table `' . $row[0] . '`');
 		}
 	}
